+++ msi.h
@@ -1,6 +1,6 @@
 #ifndef LIBRJ_MSI_H
 #define LIBRJ_MSI_H
 
-long msi(int lg, unsigned long hash, long index);
+int msi(int lg, unsigned long hash, int index);
 
 #endif
